add topnav hover states, js for dynamic styling, spacing
