+++ pgml-dashboard/static/css/scss/components/_badges.scss
@@ -13,6 +13,7 @@
 
 .github-badge {
     $color: $neon-shade-100;
+    padding: 4px;
 
     p {
         margin: 0px;
@@ -21,8 +22,9 @@
         padding: 4px;
         font-size: 0.8rem;
         font-weight: 500;
-
     }
+
+    // Add right pointing arrow 
     &::after {
         content: "";
         width: 0;
